requirements.txt
flask==2.3.2
opencv-python-headless==4.7.0.72
tensorflow-cpu==2.12.0
mediapipe==0.10.3
gunicorn==20.1.0
numpy==1.23.5